Remove the explicit type declarations for some test helpers
The declarations reference a type from hspec-wai which changed between the
version we can use with reflex-platform and the latest release.  We want to
work with both of these.  Fortunately the rest is compatible so it is enough
to just not declare the type and let ghc figure it out from the available
version of hspec-wai.